+++ src/internal/decl/cnd-handlers-decl.h
@@ -1,0 +1,5 @@
+// dots.c
+r_obj* rlang_env_dots_list(r_obj* env);
+
+static
+r_obj* hnd_call = NULL;